# contracts/openapi.yaml
...
openapi: 3.1.0
info:
  title: SoulAuth
  version: "0.1.0"
  description: >-
    Actor-native identity and authentication infrastructure.
    Endpoint surface only; payload schemas are not yet frozen.
  license:
    name: Apache-2.0
    identifier: Apache-2.0
servers:
  - url: "{APP_URL}"
    description: >-
      The public address, set by APP_URL. It is also the OIDC issuer — the two
      must match character for character.
    variables:
      APP_URL:
        default: http://localhost:8080
components:
...
  schemas:
    Error:
      type: object
      required:
      - error
      - message
      additionalProperties: true
      description: 'The single error envelope used across the whole API. `error` is a **stable machine-readable
        code** — branch on it. `message` is prose: its wording may improve at any time and must not be
        used for decisions. Individual errors carry documented additional fields.'
      properties:
        error:
          type: string
          description: Stable error code, snake_case.
          enum:
          - internal_error
          - invalid_credentials
          - email_not_verified
          - invalid_token
          - user_not_found
          - email_exists
          - username_exists
          - oauth_error
          - password_already_set
          - invalid_user_id
          - not_found
          - validation_error
          - permission_denied
          - insufficient_permissions
          - missing_permission
          - account_suspended
          - account_inactive
          - account_deleted
          - account_locked
          - rate_limit_exceeded
          - forbidden
          - bad_request
          - not_configured
          - unauthorized
          - service_unavailable
        message:
          type: string
          description: Human-readable explanation. Not part of the contract.
        required_permission:
          type: string
          description: Carried by `missing_permission` only — which permission is missing.
        locked_until_seconds:
          type: integer
          description: Carried by `account_locked` only — seconds until the lock clears.
...
    BootstrapAdminRequest:
      properties:
        email:
          type: string
        password:
          type: string
        token:
          type: string
        username:
          type: string
      required:
      - token
      - email
      - username
      - password
      type: object
    BootstrapAdminResponse:
      properties:
        email:
          type: string
        is_admin:
          type: boolean
        user_id:
          type: string
      required:
      - user_id
      - email
      - is_admin
      type: object
...
    UserResponse:
      properties:
        account_status:
          type: object
        created_at:
          type: object
        email:
          type: string
        has_password:
          type: boolean
        id:
          type: string
        is_admin:
          type: boolean
        last_login_at:
          type: object
        membership_expiry:
          type: object
        membership_level:
          type: string
        username:
          type: string
        verified:
          type: boolean
      required:
      - id
      - email
      - username
      - is_admin
      - verified
      - membership_level
      - created_at
      - has_password
      - account_status
      type: object
...
  securitySchemes:
    # 三种，不是一种。
    #
    # 把它们混成一个 `bearerAuth` 会产生具体的误导：读契约的人会以为
    # /api/oidc/userinfo 收的是登录拿到的那枚会话令牌 —— 它收的是 OIDC
    # 访问令牌，两者由不同端点签发、生命周期也不同。
    bearerAuth:
      type: http
      scheme: bearer
      description: >-
        Session token, from POST /api/auth/login (or POST
        /api/actors/authenticate for an AI actor). Used by every /api business
        endpoint.
    oidcAccessToken:
      type: http
      scheme: bearer
      description: >-
        OIDC access token, from POST /api/oidc/token. Used by
        /api/oidc/userinfo **only**. Not interchangeable with the session token
        above.
    browserSession:
      type: apiKey
      in: cookie
      name: soulauth_session
      description: >-
        Browser session cookie. /api/oidc/authorize uses it to decide whether
        this browser is logged in; its absence is a redirect to the login page,
        not an error.
...
paths:
...
  /api/auth/me:
    get:
      tags:
      - Authentication
      operationId: auth_get_current_user
      description: >-
        The account behind the presented session token.
      description_zh: >-
        返回所出示会话令牌背后的账户。
      security:
      - bearerAuth: []
      responses:
        '200':
          description: OK
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/UserResponse'
        default:
          description: Error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
...
  /api/bootstrap/admin:
    post:
      tags:
      - Bootstrap
      operationId: bootstrap_bootstrap_admin
      description: >-
        Create the first administrator using the one-time token printed at startup. Closes
        permanently once an administrator exists, and then returns the same status for a correct
        token as for a wrong one.
      description_zh: >-
        用启动时打印的一次性令牌创建第一个管理员。一旦存在管理员，这道门永久关闭，此后正确令牌与错误令牌返回同一状态码。
      security: []
      responses:
        '200':
          description: OK
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/BootstrapAdminResponse'
        default:
          description: Error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/BootstrapAdminRequest'
...